Changed the mask from BYTE to WORD, that way the mask can be 0x3ff to cover the whole io2/3 range.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@24061 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/vic20/vic20io.c
@@ -324,7 +324,7 @@ void vic20io3_store(WORD addr, BYTE value)
 
 /* ---------------------------------------------------------------------------------------------------------- */
 
-static int decodemask(BYTE mask)
+static int decodemask(WORD mask)
 {
     int len = 255;
